Restructure workflow-notifier
to allow the mock to use the real commands with authentication
