+++ workflow-notifier/.env.example
@@ -1,0 +1,5 @@
+SARA_SERVER_URL=
+TENANT_ID=
+NOTIFIER_CLIENT_ID=
+NOTIFIER_CLIENT_SECRET=
+SARA_APP_REG_SCOPE=
