+++ workflow-notifier/docker-compose.yml
@@ -1,0 +1,7 @@
+services:
+  workflow-notifier:
+    build: .
+    env_file:
+      - .env
+    ports:
+      - "8100:8100"
