--- workflow-notifier/requirements.txt
@@ -1,3 +1,0 @@
-typer~=0.14
-requests~=2.32
-msal~=1.31
